Create a workflow to called keep-active.yml
This new workflow will create a new file called counter.txt and every 59 days the workflow will run as an action. When it runs it will update the number in the counter file. By having these commits happen every couple of months this marks the repository as active and so all actions on this repository will continue to run and not get deactivated.